+++ src/modules/Playerbot/AI/ClassAI/Rogues/RogueAI.cpp
@@ -198,7 +198,7 @@ Position RogueCombatPositioning::CalculateOptimalPosition(Unit* target, RogueSpe
 
             {
 
-                float angle = target->GetOrientation() + M_PI; // 180 degrees behind
+                float angle = target->GetOrientation() + static_cast<float>(M_PI); // 180 degrees behind
                 float distance = 3.0f; // Close melee range
 
                 float x = target->GetPositionX() + distance * ::std::cos(angle);
